move version-manager to /lib
