+++ packages/usdk/lib/version-manager.mjs
